added entities in repositories

+++ src/main/java/com/example/security/user/UserEntity.java
@@ -23,8 +23,7 @@ public class UserEntity implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
+    private Integer id;
     private String firstName;
     private String lastName;
     private String email;
@@ -67,5 +66,4 @@ public class UserEntity implements UserDetails {
     public boolean isEnabled() {
         return true;
     }
-
 }
